FE: move decorate column type helper to separate file
